refactor: split news user tasks
